requirements.txt
boto3==1.24.95
hurry-filesize==0.9
psycopg==3.1.3
